Tentacle scouting work
This build has the scouting implemented, but it is broken somewhere. I removed the rotating raycasts for now, as I know those weren't working, but it seems that the normal raycasting isn't working as well. The Attack player hasn't been set up yet, but my focus was more on the scouting portion rather than the attacking portion for this sprint.